# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/GetUserScanInfo": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录token",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        },
        "/api/getcode": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取注册码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名（邮箱）",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        },
        "/api/getid": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取in3d访问id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录token",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        },
        "/api/getscan": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取in3d扫描",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录token",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "in3d访问id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        },
        "/api/getscanResult": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取in3d扫描结果",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录token",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "in3d访问id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        },
        "/api/getuserModel": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取用户模型列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录token",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "注册码",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        },
        "/api/resetPassword": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名（邮箱）",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        },
        "/api/upload": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录token",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "in3d访问id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.JSON"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "render.JSON": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务错误码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "错误描述",
                    "type": "string"
                }
            }
        }
    }
}